+++ packages/maskbook/src/components/InjectedComponents/NFT/hooks/useNFTAvatar.ts
@@ -1,0 +1,11 @@
+import { useAsync } from 'react-use'
+import { getNFTAvatar } from '../gun'
+
+export function useNFTAvatar(userId?: string) {
+    return useAsync(async () => {
+        if (!userId) return undefined
+
+        const avatar = await getNFTAvatar(userId)
+        return avatar
+    }, [userId]).value
+}
