+++ packages/maskbook/src/components/InjectedComponents/NFT/types.ts
@@ -1,0 +1,17 @@
+export interface AvatarMetaDB {
+    userId: string
+    tokenId: string
+    amount: string
+    image?: string
+    name?: string
+    address: string
+    avatarId?: string
+    symbol: string
+}
+
+export interface NFT {
+    amount: string
+    symbol: string
+    name: string
+    image: string
+}
