+++ packages/maskbook/src/social-network-adaptor/twitter.com/injection/NFT/NFTAvatarInTwitter.tsx
@@ -4,15 +4,13 @@ import { MutationObserverWatcher } from '@dimensiondev/holoflows-kit'
 import { makeStyles, useSnackbar } from '@masknet/theme'
 import { useState, useEffect, useCallback } from 'react'
 import { useCurrentVisitingIdentity } from '../../../../components/DataSource/useActivatedUI'
-import {
-    AvatarMetaDB,
-    saveNFTAvatar,
-    setOrClearAvatar,
-    useNFTAvatar,
-} from '../../../../components/InjectedComponents/NFT/NFTAvatar'
+
 import { updateAvatarImage } from '../../utils/updateAvatarImage'
 import { getAvatarId } from '../../utils/user'
 import { NFTBadge } from '../../../../components/InjectedComponents/NFT/NFTBadge'
+import { saveNFTAvatar, setOrClearAvatar } from '../../../../components/InjectedComponents/NFT/gun'
+import { useNFTAvatar } from '../../../../components/InjectedComponents/NFT/hooks'
+import type { AvatarMetaDB } from '../../../../components/InjectedComponents/NFT/types'
 
 export function injectNFTAvatarInTwitter(signal: AbortSignal) {
     const watcher = new MutationObserverWatcher(searchTwitterAvatarSelector())
@@ -68,7 +66,7 @@ function NFTAvatarInTwitter(props: NFTAvatarInTwitterProps) {
                     setAmount(avatar.amount)
                     setAvatar(avatar)
                 })
-                .catch((error) => {
+                .catch((error: Error) => {
                     enqueueSnackbar(error.message, { variant: 'error' })
                 })
         },
