+++ packages/maskbook/src/social-network-adaptor/twitter.com/injection/NFT/profileNFTAvatar.tsx
@@ -4,7 +4,8 @@ import { makeStyles } from '@masknet/theme'
 import type { ERC721TokenDetailed } from '@masknet/web3-shared'
 import { useCallback, useEffect, useState } from 'react'
 import { useMyPersonas } from '../../../../components/DataSource/useMyPersonas'
-import { NFTAvatar, useNFTAvatar } from '../../../../components/InjectedComponents/NFT/NFTAvatar'
+import { useNFTAvatar } from '../../../../components/InjectedComponents/NFT/hooks'
+import { NFTAvatar } from '../../../../components/InjectedComponents/NFT/NFTAvatar'
 import { activatedSocialNetworkUI } from '../../../../social-network'
 import { createReactRootShadowed, Flags, MaskMessage, NFTAvatarEvent, startWatch } from '../../../../utils'
 import {
